requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.81
botocore==1.35.81
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
clerk-backend-api==1.4.0
cryptography==43.0.3
Deprecated==1.2.15
Django==5.1.3
django-cors-headers==4.6.0
django-phone-verify==3.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
eval_type_backport==0.2.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonpath-python==1.0.6
multidict==6.1.0
mypy-extensions==1.0.0
nexmo==2.5.2
packaging==24.2
phonenumbers==8.13.50
pillow==11.0.0
propcache==0.2.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
twilio==9.3.7
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wrapt==1.17.0
yarl==1.18.0
